requirements.txt
torch==2.1.0+cu121
torchvision==0.16.0+cu121
PyYAML==6.0
h5py==3.10.0
tensorboard==2.10.0
packaging==23.2
scikit-image==0.22.0
six==1.16.0
tqdm==4.64.0
opencv-python==4.9.0.80
pynrrd==1.0.0